Combines all steps into a single file
guide-mixture.py performs all the steps that are necessary for running
the STAN model. The input is the matrix market file with the guide/cell/count
data. The previously manual steps that output lots of files are now python functions
and intermediate values are stored in memory. Python will handle paralellizing
the work across however many processors you have or you can set a specific number.